Migrated rescue_people to super thin templates
